Add :deps/prep-lib support for git dependencies
Source dependencies may require a "prep" step e.g. to compile Java
source[1]. This patch adds support for this but only for git
dependencies since that's the only dependency type which implements
this at the moment anyway[2]. This is achieved by extracing the
necessary info from `deps.edn` to the lockfile. A second pass over the
constructed Clojure home derived from the lockfile then picks out all
the git libs which need prepping, runs the respective command for them
and finally constructs a new home with the prepped libs.

1: See https://clojure.org/guides/deps_and_cli#prep_libs
2: Technically, :local/root dependencies also implement it but
   `clojure-nix-locker` doesn't have to handle these.