Stabilize maven-metadata.xml
This is a follow-up to https://github.com/bevuta/clojure-nix-locker/pull/14 which turned out to not
work reliably because maven-metadata.xml files may change over time as new versions of an artifact
get released. Instead, we now transform maven-metadata.xml files into a form which makes it stable
against server-side updates but still allows maven to resolve version range dependencies with
it. This transformed version is inlined with the lockfile since it's usually very small and applying
the transformation at locktime is cheaper (and easier) than doing it at use time.